Create WebUI_tesr
The idea of this is a web interface that mimics the desires for what we want to run via terminals on bare hardware but to allow the overall program to work over the internet on simple devices. The desire is to set up a server based search and a local based search.